initial project , starting again to create a simple to do lists , a new approach to finish project
