+++ src/components/AddTaskPanel/AddTaskPanel.js
@@ -1,0 +1,18 @@
+import React, { Component } from 'react';
+
+
+
+class AddTaskPanel extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {}
+    }
+    render() {
+        return (
+            <div>
+                <h2>Dodaj Task</h2>
+            </div>);
+    }
+}
+
+export default AddTaskPanel;
