+++ src/components/App/App.js
@@ -1,0 +1,23 @@
+import './App.css';
+import React from 'react'
+import AddTaskPanel from '../AddTaskPanel/AddTaskPanel';
+import TaskList from '../TaskList/TaskList';
+
+
+
+class App extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {}
+  }
+  render() {
+    return (
+      <div className='app'>
+        <AddTaskPanel />
+        <TaskList />
+      </div>
+    );
+  }
+}
+
+export default App;
