--- src/components/CountDownTimer/CountDownTimer.js
@@ -1,50 +1,0 @@
-import React from 'react'
-import NameTask from '../NameTask/NameTask'
-import './CountDown.css'
-
-class CountDownTimer extends React.Component {
-    constructor(props) {
-        super(props)
-    }
-
-    // deleting task 
-    handelDeleteTask = () => {
-        this.props.deleteTask(this.props.name, this.props.element)
-    }
-    // Handle STOP PAUSA START 
-
-    handleStopButton = () => {
-        this.props.onStopCountDown()
-    }
-
-    handleStartButton = () => {
-        this.props.onStartCountDown(this.props.index)
-        // console.log(this.props.onStartCountDown)
-    }
-    handleTogglePause = () => {
-        this.props.onTogglePause()
-    }
-
-    render() {
-        const { isPaused, minutes, seconds, isRunning, countPausa } = this.props
-        return (
-            <div>
-                <div className={isPaused ? 'inactive' : ''}>
-                    <NameTask name={this.props.name} />
-                    <p>Time Remaining: {minutes} : {seconds < 10 ? '00' : seconds}</p>
-                </div>
-
-                <button type="button" onClick={this.handleStartButton} disabled={isRunning}>Start</button>
-                <button type="button" onClick={this.handleStopButton} disabled={!isRunning} >Stop</button>
-                <button type="button" onClick={this.handleTogglePause} disabled={!isRunning}>{isPaused ? 'Continue' : 'Pause'} </button>
-                <p>Number of breaks : {countPausa}</p>
-
-                <button type="button" onClick={this.handelDeleteTask} >Delete Task</button>
-
-            </div>
-        )
-    }
-
-}
-
-export default CountDownTimer
