+++ src/components/Task/Task.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const Task = () => {
+    return (
+        <div>
+            <h4> Nasz Task</h4>
+        </div>);
+}
+
+export default Task;
